Displaying list of transaction categories. Change list by choosing different periods

+++ transactions/src/main/java/com/example/transactions/presentation/models/TransactionCategory.kt
@@ -2,5 +2,5 @@ package com.example.transactions.presentation.models
 
 data class TransactionCategory(
     val title: String,
-    val amount: Int
+    val amount: Double
 )
